+++ 30411/Baban A. Maria/4th_homework/print.h
@@ -1,0 +1,8 @@
+#ifndef PRINT_H_INCLUDED
+#define PRINT_H_INCLUDED
+#include "insert.h"
+#include "echilibrum.h"
+#include "create.h"
+void printElement(NodeT *root);
+void prettyPrint(NodeT *root,int recLevel);
+#endif //PRINT_H_INCLUDED
